// ui/src/features/rum/api/rum-journeys.generated.ts
export type RumJourneysQuery = {
  readonly tracesPage: {
    readonly total: string
    readonly items: ReadonlyArray<{
      readonly traceId: string
      readonly rootName: string
      readonly service: string
      readonly startNanos: string
      readonly durationNs: string
      readonly spanCount: number
      readonly hasError: boolean
    }>
  }
}