// ui/src/features/rum/api/rum-session.generated.ts
export type RumSessionQuery = {
  readonly rumSession: {
    readonly session: {
      readonly sessionId: string
      readonly service: string
      readonly startNanos: string
      readonly endNanos: string
      readonly spanCount: number
      readonly traceCount: number
      readonly viewCount: number
      readonly vitalCount: number
      readonly errorCount: number
      readonly hasError: boolean
    }
    readonly views: ReadonlyArray<{
      readonly tsNanos: string
      readonly screen: string
      readonly path: string | null
      readonly traceId: string
      readonly spanId: string
    }>
    readonly vitals: ReadonlyArray<{
      readonly tsNanos: string
      readonly name: string
      readonly value: number
      readonly rating: string | null
      readonly traceId: string
      readonly spanId: string
    }>
    readonly errors: ReadonlyArray<{
      readonly tsNanos: string
      readonly name: string
      readonly errorType: string | null
      readonly message: string
      readonly traceId: string
      readonly spanId: string
    }>
  } | null
}